code_deploy/requirements.txt
numpy==1.26.1
Pillow==10.1.0
pydicom==2.4.3
